ADEN-11320 fix build
